(#13, #14, #15) rework discovery
Previously in a early itteration of this client the options and discovery
would persiste between calls to Do() which while possibly handy was a bit
magical and made stats really weird - just like in the ruby client.

So we decided to ditch that but the client.Discover() remained and documented
that it will store targets and discovery time stats - while it would do none
of these things as the rest of the client have moved away from that model so
this was now just ill conceived.

We now ditch the whole client.Discover() idea and instead allow passing
a protocol.Filter into the Do() method and if that is done and no targets
are set then discovery will be done using that filter

Additionally the collective set using options will now be passed along
to the broadcast discovery system and stats handling is improved

Additionally DiscoveryTimeout can now be overriden

This is a breaking change but its for the best