(#126) ensure actions with no inputs pass validation
As per #126, currently actions with no inputs fail validation. This
is due to the fact that the `process_results` flag is added
automatically by MCO RPC ruby client. This flag is not used but
our validation code does not account for this in the
case of actions without inputs.

This commit improves the way this is ignored - we now ignore it in all
cases by simply deleting it if it is present.

This commit also adds tests to cover this new functionality.

Resolves #126.